watchdog.sh: Added ls | tail monitoring of outdir.
webcomic_walker.sh: Added MD5 sum checking as an option. If enabled, when the scraper downloads an image, it will make and store an MD5 sum of it. When downloading future images, the downloaded image is checked against the array of MD5 sums, and if a match is found, the most recently downloaded image is deleted, the count is rolled back, and the scraper moves onto the next image. Useful for comics that run reruns.